client id tag
when clients join the server, there will be a dictionary which has every client matched with the current length of the clients list. first client will have an id of 0, second will be 1. as i write this, i realize ill have to add something to make sure no id's are identical. but yea, connect 4 will drop your id number into the square now.